requirements.txt
streamlit==1.28.1
pandas==2.1.4
sqlite3
scikit-learn==1.3.2
imbalanced-learn==0.11.0
plotly==5.17.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.13
faiss-cpu==1.7.4
python-dotenv==1.0.0